Remove mixed architecture example
Mixed architectures are not possible in SVD by definition. The idea
there is that if there are two processors, a 64-bit one and a 32-bit
one, they will both have their own SVD file. In this case, some devices
might use the 64-bit address space while others might only use the
32-bit address space.

At the end of the day we are interested in two things:

1. What is the target CPU architecure, 64-bit or 32-bit?
2. What is the width of the target register: 8-, 16-, 32-, 64-bit, etc.

How a target address is generated from a system description is not
trivial. It depends both on bus layout and peripheral register maps.